Temporary glslang warning fix

+++ renderdoc/3rdparty/glslang/glslang/MachineIndependent/Scan.h
@@ -172,7 +172,7 @@ public:
 
     void setFile(const char* filename, size_t i)
     {
-        if (i == getLastValidSourceIndex()) {
+        if ((int)i == getLastValidSourceIndex()) {
             logicalSourceLoc.name = filename;
         }
         loc[i].name = filename;
